requirements.txt
fastapi
uvicorn
requests
pillow
edge-tts
python-multipart
